+++ module/helpers/templates.mjs
@@ -21,6 +21,7 @@ export const preloadHandlebarsTemplates = async function () {
     "systems/mist-engine-fvtt/templates/actor/parts/character-quintessences-partial.hbs",
     "systems/mist-engine-fvtt/templates/actor/parts/actor-fellowship-partial.hbs",
     "systems/mist-engine-fvtt/templates/actor/parts/actor-fellowship-themebook-partial.hbs",
+    "systems/mist-engine-fvtt/templates/actor/parts/challenge-partial.hbs",
     "systems/mist-engine-fvtt/templates/actor/parts/floating-tags-and-status-partial.hbs",
     "systems/mist-engine-fvtt/templates/shared/lock-toggle.hbs",
     "systems/mist-engine-fvtt/templates/shared/floating-tags-and-status-partial.hbs",
@@ -29,6 +30,7 @@ export const preloadHandlebarsTemplates = async function () {
     "systems/mist-engine-fvtt/templates/actor/litm-fellowship-themecard/tab-themebook.hbs",
     "systems/mist-engine-fvtt/templates/item/parts/themebook-special-improvement-line.hbs",
     "systems/mist-engine-fvtt/templates/actor/litm-fellowship-themecard/themebook-special-improvements-line.hbs",
-    "systems/mist-engine-fvtt/templates/scene-app/_single_character.hbs"
+    "systems/mist-engine-fvtt/templates/scene-app/_single_character.hbs",
+    "systems/mist-engine-fvtt/templates/actor/parts/npc-challenges-edit-partial.hbs"
   ]);
 };