+++ module/lib/tag-status-text-helper.mjs
@@ -1,24 +1,45 @@
 export function makeStyledTagOrStatusText(source){
     let isStatus = false;
+    let extraClass = '';
+    let extraIcon = '';
+
+    // status
     if(source.includes("/s")){
         isStatus = true;
-        source = source.replace("/s", "");
+        
+        if(source.includes("/sg")){
+          extraClass = "green";
+          source = source.replace("/sg", "");
+        }else{
+          source = source.replace("/s", "");
+        }
+    }
+
+    // weakness
+    if(source.includes("/w")){
+      extraIcon = '<i class="fa-light fa-angles-down"></i>';
+      if(source.includes("/wo")){
+          extraClass = "orange";
+          source = source.replace("/wo", "");
+        }else{
+          source = source.replace("/w", "");
+        }
     }
 
     if(!source.includes("-")){
         let tag = source;
         if(isStatus){
-            return `<mark class="draggable status" draggable="true" data-type="status" data-name="${tag}" data-value="0">${tag}</mark>`
+            return `<mark class="draggable status ${extraClass}" draggable="true" data-type="status" data-name="${tag}" data-value="0">${extraIcon}${tag}</mark>`;
         }else{
-            return `<mark class="draggable tag" draggable="true" data-type="tag" data-name="${tag}">${tag}</mark>`
+            return `<mark class="draggable tag ${extraClass}" draggable="true" data-type="tag" data-name="${tag}">${extraIcon}${tag}</mark>`;
         }
     }else{
         const value = source.split("-").pop();
         const name = source.substring(0, source.lastIndexOf("-"));
-        return `<mark class="draggable status" draggable="true" data-type="status" data-name="${name}" data-value="${value}">${name}-${value}</mark>`;
+        return `<mark class="draggable status ${extraClass}" draggable="true" data-type="status" data-name="${name}" data-value="${value}">${name}-${value}</mark>`;
     }
 
-    const [name, value] = source.split("-");
+    /*const [name, value] = source.split("-");
     console.log("name:", name, "value:", value);
     return;
     if (tag.includes("-")) {
@@ -31,5 +52,5 @@ export function makeStyledTagOrStatusText(source){
         `[${tag}]`,
         `<mark class="draggable tag" draggable="true" data-type="tag" data-name="${name}">${tag}</mark>`
       );
-    }
+    }*/
 }